+++ extensions/instagram-media-downloader/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Instagram Media Downloader Changelog
 
+## [Update] - {PR_MERGE_DATE}
+
+- Update the method for obtaining the Instagram media URL.
+
 ## [Initial Version] - 2025-05-21